RedditStream/LandingZone/requirements.txt
Flask==2.1.1
Werkzeug==2.0.3
requests==2.28.1
google-cloud-storage==2.5.0
